% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{kest.gt}
\alias{kest.gt}
\title{Expected number of classes with more than N individuals given parameters \eqn{\theta} and sample size J using the species frequency spectrum}
\usage{
kest.gt(theta, J, N)
}
\arguments{
\item{theta}{parameter \eqn{\theta}, a positive real value}

\item{J}{number of individuals, a whole number}

\item{N}{minimal abundance, with \verb{0 <= N < J}}
}
\value{
An object of class \code{"integrate"} as returned by \code{\link[stats:integrate]{stats::integrate()}}:
the expected number of classes \eqn{>N} is its \verb{$value} component, and
\verb{$abs.error} is the estimated absolute error of the numerical integration.
}
\description{
The expected number of species \eqn{>N} is computed using the approximation
based on the species frequency spectrum \eqn{\psi(x)=\theta x^{-1}(1-x)^{\theta-1}}
such that the number of species \eqn{>N} is the integral from N/J to 1 of \eqn{(1-(1-x)^J)\psi(x)}.
This is an approximation of the exact formula, but it turns out to be accurate.
}
\examples{
kest.gt(11.3, 234373, 50)$value
}
