% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutr-package.R
\docType{package}
\name{neutr-package}
\alias{neutr}
\alias{neutr-package}
\title{neutr: Inference and Fast Generation of Neutral Species Assemblages}
\description{
Parameter inference and fast generation for three classes of neutral species
assemblage models: the Ewens model, the multideme model and the Pitman model.
}
\section{Estimation}{

Parameters are estimated by maximising the likelihood: \code{\link[=optim.ewens]{optim.ewens()}} for the
Ewens model, \code{\link[=optim.pitman]{optim.pitman()}} for the Pitman model and \code{\link[=optim.multideme]{optim.multideme()}} for
the multideme model.
}

\section{Generation}{

\code{\link[=generate.hoppe.urn0]{generate.hoppe.urn0()}} simulates Hoppe's urn exactly, individual by individual;
\code{\link[=generate.hoppe.urn]{generate.hoppe.urn()}} and \code{\link[=generate.pitman.urn]{generate.pitman.urn()}} use the
Griffiths-Engen-McCloskey representation, whose cost does not depend on the
sample size, and scale to 10^12 individuals and beyond.
}

\section{Expected richness}{

\code{\link[=kest]{kest()}}, \code{\link[=kest.gt]{kest.gt()}} and \code{\link[=kest.gt0]{kest.gt0()}} give the expected number of species
(possibly above an abundance threshold).
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeromechave/neutr}
  \item Report bugs at \url{https://github.com/jeromechave/neutr/issues}
}

}
\author{
Jerome Chave
}
\keyword{internal}
