% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zeroremove}
\alias{zeroremove}
\title{General use function}
\usage{
zeroremove(x)
}
\arguments{
\item{x}{vector of real}
}
\value{
the vector \code{x} without its zero and missing (\code{NA}/\code{NaN}) elements
}
\description{
\code{zeroremove()} is used to remove zeros and NAs in a vector of reals
or of integers. This is a simple utility function.
}
\examples{
zeroremove(c(2.1, 0.0, 3.0, NA))
}
