% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutr.R
\name{generate.hoppe.urn}
\alias{generate.hoppe.urn}
\title{generate.hoppe.urn}
\usage{
generate.hoppe.urn(theta, J)
}
\arguments{
\item{theta}{real}

\item{J}{integer}
}
\value{
output_abundances a vector of integers

k a real number
}
\description{
generate.hoppe.urn
}
\examples{
generate.hoppe.urn(11.3,234373)
}
