% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutr.R
\name{generate.hoppe.urn}
\alias{generate.hoppe.urn}
\title{Generation of a neutral partition given parameter \eqn{\theta} using the Griffith-Engen-McCloskey representation}
\usage{
generate.hoppe.urn(theta, J)
}
\arguments{
\item{theta}{parameter \eqn{\theta}, a real value}

\item{J}{number of individuals, an integer}
}
\value{
output_abundances is a vector of integers

k is the expected number of classes, a real number
}
\description{
generate.hoppe.urn creates a rank-abundance distribution for a given \eqn{\theta} and J.
Beta-distributed random variables are picked from distribution \eqn{Beta(1,\theta)},
and the ressulting partition is one multinomial draw of J individuals drawn according to the GEM random variables.
}
\examples{
generate.hoppe.urn(11.3,234373)
}
