% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutr.R
\name{generate.pitman.urn}
\alias{generate.pitman.urn}
\title{generate.pitman.urn}
\usage{
generate.pitman.urn(theta, sigma, J)
}
\arguments{
\item{theta}{real}

\item{sigma}{real}

\item{J}{integer}
}
\value{
output_abundances a vector of integers

k a real number
}
\description{
generate.pitman.urn
}
\examples{
generate.pitman.urn(11.3,0.1,234373)
}
