% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutr.R
\name{kest.gt}
\alias{kest.gt}
\title{Expected number of classes with more than N individuals given parameters \eqn{\theta} and sample size J for the Hoppe urn model}
\usage{
kest.gt(theta, J, N)
}
\arguments{
\item{theta}{parameter \eqn{\theta},  a real value}

\item{J}{is the number of individuals, an integer}

\item{N}{is the minimal abundance, an integer}
}
\value{
k is the expected number of classes \eqn{>N}, a real number
}
\description{
The expected number of species \eqn{>N} is computed using 100 runs of
the Hoppe urn scheme
}
\examples{
kest.gt(11.3,234373,50)
}
