% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutr.R
\name{kest}
\alias{kest}
\title{Expected number of classes given parameters \eqn{\theta} and sample size J for the Hoppe urn model}
\usage{
kest(theta, J)
}
\arguments{
\item{theta}{parameter \eqn{\theta},  a real value}

\item{J}{is the number of individuals, an integer}
}
\value{
k is the expected number of classes, a real number
}
\description{
The expected number of species is computed using an exact expression
based on the diagamma function.
}
\examples{
kest(11.3,234373)
}
