% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutr.R
\name{optim.ewens}
\alias{optim.ewens}
\title{optim.ewens}
\usage{
optim.ewens(input_abundances)
}
\arguments{
\item{input_abundances}{a vector of integers}
}
\value{
theta a real value

log-likelihood a real value
}
\description{
optim.ewens
}
\examples{
input_abundances=c(234,87,34,5,4,3,3,2,2,1,1,1,1)
optim.ewens(input_abundances)
}
