% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutr.R
\name{optim.ewens}
\alias{optim.ewens}
\title{Maximum likelihood estimation of theta based on Ewens sampling formula}
\usage{
optim.ewens(input_abundances)
}
\arguments{
\item{input_abundances}{a vector of integers}
}
\value{
theta parameter of Ewens sampling formula, a real value

logl maximal value of the log-likelihood, a real value
}
\description{
optim.ewens computes the maximum likelihood estimate of Ewens theta parameter
for a given species abundance distribution {n_1, ..., n_k}
where n_i is the number of organisms of species i in the sample
and k is the total number of species (for all i, n_i > 0)
The input should be a vector
}
\examples{
input_abundances=c(234,87,34,5,4,3,3,2,2,1,1,1,1)
optim.ewens(input_abundances)
}
