% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutr.R
\name{optim.multideme}
\alias{optim.multideme}
\title{Maximum likelihood estimation of local immigration rates \eqn{(m_1,...,m_K)} based on the K-deme sampling formula}
\usage{
optim.multideme(input_abundance_matrix)
}
\arguments{
\item{input_abundance_matrix}{a matrix \eqn{n_{ij}} of abundances for species i in deme (local site) j}
}
\value{
I rescaled immigration rate, \eqn{I=\frac{m}{1-m}(J-1)}, a vector of real numbers

m immigration rate, a vector of real numbers

J local community size, a vector of integers

k number of classes/species, a vector of integers
}
\description{
optim.multideme computes the maximum likelihood estimate for the multi-deme parameters
for a given species abundance matrix \eqn{(n_{1j}, ..., n_{kj})} for deme j (\eqn{j\in\{1,...,K\}})
where \eqn{n_{ij}} is the number of organisms of species i in deme j
and k is the total number of species (for all \eqn{i, n_i \geq 0}).
The formulation in this description is that of species abundance distributions
but is valid for any partition in a subdivided setting (e.g., word frequencies in multiple books).
}
\examples{
input_abundances1=c(44,37,34,5,4,3,3,2,2,1,1,1,1)
input_abundances2=c(240,20,48,2,21,1,3,2,5,2,0,1,1)
input_abundance_matrix=rbind(input_abundances1,input_abundances2)
optim.multideme(input_abundance_matrix)
}
