% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutr.R
\name{optim.multideme}
\alias{optim.multideme}
\title{optim.multideme}
\usage{
optim.multideme(input_abundance_matrix)
}
\arguments{
\item{input_abundance_matrix}{a matrix of abundances}
}
\value{
I a vector of real numbers

m a vector of real numbers

J a vector of integers

k a vector of real numbers
}
\description{
optim.multideme
}
\examples{
input_abundances1=c(234,87,34,5,4,3,3,2,2,1,1,1,1)
input_abundances2=c(240,20,48,2,21,1,3,2,5,2,1,1,1)
input_abundance_matrix=rbind(input_abundances1,input_abundances2)
optim.multideme(input_abundance_matrix)
}
