% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutr.R
\name{optim.pitman}
\alias{optim.pitman}
\title{Maximum likelihood estimation of \eqn{\theta} and \eqn{\sigma} based on Pitman sampling formula}
\usage{
optim.pitman(input_abundances, init_vals = c(10, 0.1))
}
\arguments{
\item{input_abundances}{a vector of integers}

\item{init_vals}{initial values of (theta,sigma), a vector of two real numbers}
}
\value{
theta parameter \eqn{\theta}, a real value

sigma parameter \eqn{\sigma}, a real value

logl, maximal log-likelihood, a real value
}
\description{
optim.pitman computes the maximum likelihood estimate of Pitman \eqn{(\theta,\sigma)} parameters
for a given species abundance distribution \eqn{(n_1, ..., n_k)}
where \eqn{n_i} is the number of organisms of species i in the sample
and k is the total number of species (for all \eqn{i, n_i > 0})
The input should be a vector and initial values of \eqn{(\theta,\sigma)}
with the argument init_vals=c(theta_init,sigma_init)
with the condition theta_init>0.0, and 1.0 > sigma_init > 0.0.
}
\examples{
optim.pitman(c(234,87,34,5,4,3,3,2,2,1,1,1,1),c(10.0,0.1))
}
