% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutr.R
\name{optim.pitman}
\alias{optim.pitman}
\title{optim.pitman}
\usage{
optim.pitman(input_abundances, init_vals = c(10, 0.1))
}
\arguments{
\item{input_abundances}{a vector of integers}

\item{init_vals}{a vector of two real numbers}
}
\value{
theta a real value

sigma a real value

log-likelihood a real value
}
\description{
optim.pitman
}
\examples{
optim.pitman(c(234,87,34,5,4,3,3,2,2,1,1,1,1),c(10.0,0.1))
}
